+++ src/store/actions/getCurrentWeatherFail.ts
@@ -1,0 +1,7 @@
+import { ISetErrorAction, SET_ERROR } from "../types/currentWeather";
+
+export const getWeatherFail = (payload: string): ISetErrorAction => ({
+    type: SET_ERROR,
+    payload,
+  });
+
