+++ src/store/actions/getCurrentWeatherPending.ts
@@ -1,0 +1,6 @@
+import { ISetLoadingAction, SET_LOADING } from "../types/currentWeather";
+
+export const getWeatherPending = (): ISetLoadingAction => ({
+    type: SET_LOADING,
+  });
+
