+++ src/store/actions/getCurrentWeatherSuccess.ts
@@ -1,0 +1,12 @@
+import {
+  GET_WEATHER_SUCCESS,
+  IGetWeatherAction,
+  IWeatherData,
+} from "../types/currentWeather";
+
+export const getWeatherSuccess = (
+  payload: IWeatherData
+): IGetWeatherAction => ({
+  type: GET_WEATHER_SUCCESS,
+  payload,
+});
