+++ src/store/actions/getWeatherDailySucces.ts
@@ -1,0 +1,6 @@
+import { GET_DAILY_WEATHER_SUCCESS, IGetDailyWeatherAction, IWeatherDailyData } from "../types/dailyWeather";
+
+export const getWeatherDailySuccess = (payload: IWeatherDailyData): IGetDailyWeatherAction => ({
+    type: GET_DAILY_WEATHER_SUCCESS,
+    payload,
+  });
